Fix build on osx

+++ fsutil/statfs_unix.go
@@ -13,7 +13,7 @@ func Statfs(path string) (FsStat, error) {
 	}
 
 	return FsStat{
-		Capacity:  int64(stat.Blocks) * stat.Bsize,
-		Available: int64(stat.Bavail) * stat.Bsize,
+		Capacity:  int64(stat.Blocks) * int64(stat.Bsize),
+		Available: int64(stat.Bavail) * int64(stat.Bsize),
 	}, nil
 }
